pyproject.toml
[project]
name = "glucose-neuralforecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "eliot>=1.17.5",
    "kaleido>=0.2.1",
    "matplotlib>=3.10.7",
    "neuralforecast>=3.1.2",
    "plotly>=6.3.1",
    "polars>=1.34.0",
    "pycomfort>=0.0.18",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "typer>=0.20.0",
    "utilsforecast>=0.2.14",
]

[project.scripts]
train = "glucose_neuralforecast.train:app"
predict = "glucose_neuralforecast.inference:app"
summarize = "glucose_neuralforecast.summarize:app"
list-models = "glucose_neuralforecast.train:app"
list-trained = "glucose_neuralforecast.inference:app"
generate-config = "glucose_neuralforecast.train:app"
train-from-config = "glucose_neuralforecast.train:app"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "jupyterlab>=4.4.9",
]
